+++ CHANGES.rst
@@ -1,3 +1,11 @@
+0.2 (2019)
++++++++++++++++++++++++
+
+- Paper release
+- Updated all example notebooks
+- Include subdividing model (Eddy)
+
+
 0.1 (2019)
 +++++++++++++++++++++++